Move searchfrom into tipsyness subdirectory
